% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterpp_gini.R
\name{clusterpp_gini}
\alias{clusterpp_gini}
\title{Gini on point clusters}
\usage{
clusterpp_gini(sfcluster)
}
\arguments{
\item{sfcluster}{A sf point patterns and their belonging to clusters}
}
\value{
Gini index
}
\description{
Compute gini index on clusters of spatial points
}
\seealso{
\code{\link[=dist_clustering]{dist_clustering()}} for clustering points on planar networks
}
