% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_2_pp.R
\name{dist_2_pp}
\alias{dist_2_pp}
\title{Shortest paths distances between two points pattern on a planar network}
\usage{
dist_2_pp(pp1, pp2, network)
}
\arguments{
\item{pp1}{A sf point patterns}

\item{pp2}{A sf point patterns}

\item{network}{A sf planar networks}
}
\value{
A tibble of shortest paths
}
\description{
Computation of shortest path pairwise distances between between two point patterns of a planar network.
Input point patterns data must be snapped on linear networks.
}
\details{
\itemize{
\item \code{pp1}/\code{pp2} points numbers in output tibble refer to initial orders of rows in point patterns data.
}
}
\seealso{
\code{\link[=snapPointsToLines]{snapPointsToLines()}} for snapping points on lines.
}
