% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cluster.R
\name{dist_clustering}
\alias{dist_clustering}
\title{Clustering points on planar networks}
\usage{
dist_clustering(pp, network, distparam)
}
\arguments{
\item{pp}{A sf point patterns}

\item{network}{A sf planar networks}

\item{distparam}{The distance chosen for clustering}
}
\value{
The sf point patterns and their belonging to clusters
}
\description{
Compute shortest path distances between all observed points on a planar network and
clustering the points with HCA clustering method (average parameter). Final clusters are
based on a chosen distance parameter applied to the dendrogram result of HCA.
Input point patterns data must be snapped on linear networks.
}
\seealso{
\itemize{
\item \code{\link[=snapPointsToLines]{snapPointsToLines()}} for snapping points on lines
\item \code{\link[=hclust]{hclust()}} and \code{\link[=cutree]{cutree()}} for HCA and clustering methods
}
}
