% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clustering_list.R
\name{dist_clustering_list}
\alias{dist_clustering_list}
\title{Clustering points (class list) on planar networks (class list)}
\usage{
dist_clustering_list(listpp, listnetwork, distparam)
}
\arguments{
\item{listpp}{A list of sf point patterns}

\item{listnetwork}{A list of sf planar networks}

\item{distparam}{The distance chosen for clustering}
}
\value{
The list of sf point patterns and their belonging to clusters
}
\description{
Compute shortest path distances between all observed points of planar networks and
clustering the points with HCA clustering method (average parameter). Final clusters are
based on a chosen distance parameter applied to the dendrogram result of HCA.
Input point patterns data must be snapped on linear networks.
The length and order of lists of points and planar networks must be equivalent and consistent.
}
\seealso{
\itemize{
\item \code{\link[=snapPointsToLines]{snapPointsToLines()}} for snapping points on lines
\item \code{\link[=hclust]{hclust()}} and \code{\link[=cutree]{cutree()}} for HCA and clustering methods
}
}
