% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_with_sims.R
\name{dist_with_sims}
\alias{dist_with_sims}
\title{Shortest paths distances between points on a planar network}
\usage{
dist_with_sims(pp, network, nsim)
}
\arguments{
\item{pp}{A sf point patterns}

\item{network}{A sf planar networks}

\item{nsim}{The number of simulations generating randomly simulated points (must be > 1)}
}
\value{
A tibble of shortest paths
}
\description{
Computation of shortest path distances between all observed points of a planar network and
shortest paths distances between randomly simulated points on the same planar network.
Input point patterns data must be snapped on linear networks.
}
\details{
\itemize{
\item \code{Pi}/\code{P} points numbers in output tibble refer to initial order of rows in point patterns data.
\item \code{sim} column refers to the to the number of a specific simulation and the
\code{type} to the calculated distance depending on whether the points are simulated or observed.
}
}
\examples{
library(sf)
library(ggplot2)
pharmacy # point pattern data
paris_network # planar network data 
distances <- dist_with_sims(pp = pharmacy, network = paris_network, nsim = 2)
ggplot(data = distances, mapping = aes(x = dist_pi_p, color = type)) +
geom_density()

}
\seealso{
\code{\link[=snapPointsToLines]{snapPointsToLines()}} for snapping points on lines.
}
