% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_with_sims_list.R
\name{dist_with_sims_list}
\alias{dist_with_sims_list}
\title{Shortest paths distances between points (class list) on planar networks (class list)}
\usage{
dist_with_sims_list(listpp, listnetwork, nsim)
}
\arguments{
\item{listpp}{A list of sf point patterns}

\item{listnetwork}{A list of sf planar networks}

\item{nsim}{The number of simulations generating randomly simulated points (must be > 1)}
}
\value{
A list of tibble of shortest paths
}
\description{
Computation of shortest path distances between all observed points of planar networks and
shortest paths distances between randomly simulated points on the same planar networks.
Input point patterns data must be snapped on linear networks.
The length and order of lists of points and planar networks must be equivalent and consistent.
}
\details{
\itemize{
\item \code{Pi}/\code{P} points numbers in output tibbles refer to initial order of rows in point patterns data.
\item \code{sim} column refers to the to the number of a specific simulation and the
\code{type} to the calculated distance depending on whether the points are simulated or observed.
}
}
\seealso{
\code{\link[=snapPointsToLines]{snapPointsToLines()}} for snapping points on lines.
}
