% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_on_network.R
\name{list_dist_with_sims}
\alias{list_dist_with_sims}
\title{List of shortest path distances between points in a planar network}
\usage{
list_dist_with_sims(listpp, listnetwork, nsim)
}
\arguments{
\item{listpp}{The list of sf point patterns}

\item{listnetwork}{The list of sf planar networks lines}

\item{nsim}{The number of simulations}
}
\value{
The list of tibble of shortest paths
}
\description{
Computation of shortest path distances between all observed points of a planar network and randomly simulated points
}
\examples{
blob
}
\author{
boubou
}
