% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paris_network}
\alias{paris_network}
\title{Pharmacy and coffee data}
\format{
\subsection{\code{paris_network}}{

A simple feature with 6710 features and 4 fields:
\describe{
\item{full_id, osm_id}{Open Street Map identifiers}
\item{osm_type, highway}{Open Street Map keys}
\item{geometry}{Open Street Map pharmacy point}
}
}
}
\source{
Overpass-API \url{http://github.com/drolbr/Overpass-API}
}
\usage{
paris_network
}
\description{
Pharmacies and coffees in the central boroughs of Paris
}
\keyword{datasets}
