% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pharmacy}
\alias{pharmacy}
\title{Pharmacy data}
\format{
\subsection{\code{pharmacy}}{

A simple feature with 138 features and 2 fields:
\describe{
\item{osm_id}{Open Street Map identifiers}
\item{amenity}{Open Street Map amenity key}
\item{geometry}{Open Street Map pharmacy point}
}
}
}
\source{
Overpass-API \url{http://github.com/drolbr/Overpass-API}
}
\usage{
pharmacy
}
\description{
Pharmacies in the central boroughs of Paris
}
\keyword{datasets}
