% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pharmacy_coffee}
\alias{pharmacy_coffee}
\title{Pharmacy and coffee data}
\format{
\subsection{\code{pharmacy_coffee}}{

A simple feature with 843 features and 2 fields:
\describe{
\item{osm_id}{Open Street Map identifiers}
\item{amenity}{Open Street Map amenity key}
\item{geometry}{Open Street Map pharmacy point}
}
}
}
\source{
Overpass-API \url{http://github.com/drolbr/Overpass-API}
}
\usage{
pharmacy_coffee
}
\description{
Pharmacies and coffees in the central boroughs of Paris
}
\keyword{datasets}
