% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_on_network.R
\name{dist_with_sims_list}
\alias{dist_with_sims_list}
\title{Shortest path distances between a list of points on a list of planar network}
\usage{
dist_with_sims_list(listpp, listnetwork, nsim)
}
\arguments{
\item{listpp}{a list of sf point patterns}

\item{listnetwork}{a list of sf planar networks lines}

\item{nsim}{the number of simulations (must be > 1)}
}
\value{
a list of tibble of shortest paths
}
\description{
Computation of shortest path distances between all observed points of a planar network and randomly simulated points.
Input data point must be snapped on linear network.
The length and order of lists of points and networks must be equivalent and consistent.
}
\examples{
blob
}
