% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_on_network.R
\name{dist_with_sims_list}
\alias{dist_with_sims_list}
\title{Shortest paths distances between points (in class lists) on planar networks (in class lists)}
\usage{
dist_with_sims_list(listpp, listnetwork, nsim)
}
\arguments{
\item{listpp}{A list of sf point patterns}

\item{listnetwork}{A list of sf planar networks lines}

\item{nsim}{The number of simulations chosen for randomly simulated points (must be > 1)}
}
\value{
A list of tibble of shortest paths
}
\description{
Computation of shortest path distances between all observed points of a planar networks and
shortest paths distances between randomly simulated points on the same planar networks.
Input data point patterns must be snapped on linear networks.
The length and order of lists of points and planar networks must be equivalent and consistent.
}
\details{
Pi/P points numbers in output tibbles refer to initial order of rows in point patterns data.
}
\examples{
blob

}
\seealso{
Snapping points on lines: \code{\link[maptools::snapPointsToLines]{maptools::snapPointsToLines()}}
}
